Rename MLIR's support library to fix link issues
Even though Bazel's own rules_cc can handle this fine (by passing the
linker either a list of object files or by passing it the full paths
of the libraries rather than using `-l` syntax with `-L`), not
everything can. rules_rust cannot, for example, because it uses `rustc`
to drive the compilation of binaries which cannot correctly deal with
multiple libraries with the same name.

I left a copy of the `Support` library in here so code doesn't break;
ideally it would really be a copy or a symlink instead of another
thing to build.